perf: Verify file exists before unlink
Failures during init mean the cleanup function can be called before
the file has been opened.

+++ tools/perf/util/data.c
@@ -127,5 +127,6 @@ ssize_t perf_data_file__write(struct perf_data_file *file,
 
 void perf_data_file__unlink(struct perf_data_file *file)
 {
-	unlink(file->path);
+	if (file)
+		unlink(file->path);
 }
